requirements.txt
google-genai>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
vaderSentiment>=3.3.2
yfinance>=0.2.30
python-dotenv>=1.0.0
streamlit>=1.30.0
transformers>=4.36.0
torch>=2.1.0
